--- internal/errs/centralized.go
@@ -1,5 +1,0 @@
-package errs
-
-type ErrNoProject struct {
-	*WrapperError
-}
