+++ internal/runbits/errors/centralized.go
@@ -1,0 +1,7 @@
+package errors
+
+import "github.com/ActiveState/cli/internal/errs"
+
+type ErrNoProject struct {
+	*errs.WrapperError
+}
